Change math to 2D/3D variant, change doubles->floats

+++ src/aggregates/kdtree.h
@@ -9,7 +9,7 @@ struct KDTreeNode {
   int primOff;    // Primitives offset
   int numPrims;   // 0 -> internal node
   int axis;
-  double splitPos;
+  float splitPos;
   KDTreeNode *a, *b;
 };
