+++ src/materials/orennayar.h
@@ -7,11 +7,11 @@
 #include <core/bsdf.h>
 
 struct OrenNayar : BSDF {
-  OrenNayar(double roughness, Colour col);
+  OrenNayar(float roughness, Colour col);
   Colour eval(HitRec& rec);
   Colour sample(HitRec& rec, RNG& rng);
-  double pdf(HitRec& rec);
+  float pdf(HitRec& rec);
 
-  double A;
-  double B;
+  float A;
+  float B;
 };
