+++ src/objects/plane.cc
@@ -2,7 +2,7 @@
 
 bool Plane::hit(Ray& r, HitRec& rec) {
   Ray transformed = rayTransform(r);
-  double t = -transformed.p.z / transformed.d.z;
+  float t = -transformed.p.z / transformed.d.z;
   Vec3 it = transformed.at(t);
   if (t < TOL || t > r.tMax || 
       fabs(it.x) > 1 || fabs(it.y) > 1) {
@@ -22,9 +22,9 @@ bool Plane::hit(Ray& r, HitRec& rec) {
   }
 }
 
-Vec3 Plane::sample(double *pdf, RNG& rng) {
-  double x = rng.rand01() * 2 - 1;
-  double y = rng.rand01() * 2 - 1;
+Vec3 Plane::sample(float *pdf, RNG& rng) {
+  float x = rng.rand01() * 2 - 1;
+  float y = rng.rand01() * 2 - 1;
   *pdf = 1 / surfaceArea;
   return T * Vec3(x, y, 0);
 }
