+++ src/scenes/dof_scene.cc
@@ -10,8 +10,8 @@ SCENE(DOF) {
   Vec3 g = Vec3(0, -2, 0) - e;
   Vec3 up = Vec3(0, 1, 0);
 
-  params.set<double>("aperture", 1);
-  params.set<double>("focus_dist", 15);
+  params.set<float>("aperture", 1);
+  params.set<float>("focus_dist", 15);
 
   scene->cam = Camera(e, g, up, 70, params);
 
@@ -84,8 +84,8 @@ SCENE(DOF) {
 
       if (i == 1 && j == 0) continue;
 
-      double xOff = -6.5 + i * 6.5;
-      double zOff =    1 + j * 4;
+      float xOff = -6.5 + i * 6.5;
+      float zOff =    1 + j * 4;
 
       s = new Cylinder(new Lambertian(1));
       s->Scale(2, 1, 1);
