+++ src/util/map_converter.h
@@ -3,5 +3,5 @@
 #include <util/image.h>
 #include <core/math.h>
 
-Image convertHeightMapToNormalMap(Image& heightIm, double strength=1);
+Image convertHeightMapToNormalMap(Image& heightIm, float strength=1);
